PR changes, embedded acceptance test file structure, serverless sqs/sns deployment
