+++ backend/endpoints/sendReportSubmittedEmail/lambda.ts
@@ -1,0 +1,11 @@
+import { SQSEvent, SQSHandler } from "aws-lambda";
+
+export const handler: SQSHandler = async (event: SQSEvent) => {
+
+    try{
+        console.log("email gets sent here")
+    }
+    catch{
+        console.log("email doesn't get sent here")
+    }
+}
