+++ backend/package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@aws-sdk/client-dynamodb": "^3.131.0",
     "@aws-sdk/client-s3": "^3.32.0",
+    "@aws-sdk/client-sns": "^3.171.0",
     "@aws-sdk/lib-dynamodb": "^3.32.0",
     "@aws-sdk/node-http-handler": "^3.32.0",
     "@aws-sdk/types": "^3.32.0",
@@ -31,8 +32,10 @@
     "uuid": "^8.3.2"
   },
   "devDependencies": {
+    "@agiledigital/serverless-sns-sqs-lambda": "^2.0.1",
     "@types/aws-lambda": "^8.10.101",
     "@types/node": "^18.0.6",
+    "serverless-domain-manager": "^6.1.0",
     "serverless-offline": "^9.1.3",
     "serverless-plugin-typescript": "^2.1.2",
     "typescript": "^4.7.4"
