+++ backend/scripts/populateDatabase.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs';
 
 import { ChargeDevice } from "../types/ChargeDevice";
-import ChargeDeviceClient from "../dynamo/ChargeDeviceClient";
+import ChargeDeviceClient from "../aws/dynamo/ChargeDeviceClient";
 
 const data: ChargeDevice[] = JSON.parse(fs.readFileSync('./json/lowerCaseChargeDevice.json', 'utf8'));
