+++ frontend/src/Components/Layout/Layout.module.css
@@ -37,13 +37,6 @@ Footer {
     margin-left:5%;
 }
 
-/* .profilelink {
-    text-decoration: none;
-    color: white;
-    background: #8f1aa7;
-    display:block;
-} */
-
 .signoutbutton {
     border-right: auto;
 }
